requirements.txt
google-api-core==2.17.1
ecdsa>=0.18.0
requests>=2.31.0
jsonify>=0.5
immudb-py>=1.4.0
grpcio>=1.48.1
protobuf>=3.12.0<3.20.x
google-api==0.1.12
googleapis-common-protos==1.56.4
fastapi==0.104.1
pydantic==2.4.2
pyotp==2.9.0
starlette~=0.27.0
uvicorn==0.25.0
immudb-py==1.4.0
python-dotenv~=1.0.0
cryptography~=41.0.7
pygments>=2.17.2
fastapi >= 0.104.1
guess-language-spirit>=0.5.3
python_multipart >=0.0.9
transformers >= 4.38.2
torch >=2.2.1
petals >= 2.2.0